Separate data into Map model
